InputMananger: introduce platform-independent KeyCode abstraction and normalization layer
Modifies;
- TUI.vcxproj
- Input/
     - InputManager.cpp
     - KeyEvent.h
     - RawKeyboardSource.h

Adds:
- Input/KeyCodes.h

Establishes a stable, platform-agnostic keyboard vocabulary by introducing
KeyCode and updating the input pipeline to normalize raw platform input into this
internal representation.

Key changes:
- Added Input/KeyCodes.h:
  - Defines KeyCode enum as the foundational key vocabulary
  - Covers control keys (Enter, Escape, Tab, Backtab, Backspace, Space)
  - Covers navigation keys (arrows, Home/End, PageUp/PageDown, Insert/Delete)
  - Adds F1–F12 for extensibility (help/debug tooling)
  - Introduces KeyModifiers (shift/ctrl/alt)

- Updated KeyEvent:
  - Now uses KeyCode instead of raw platform values
  - Stores printable input via KeyCode::Character + char32_t character
  - Preserves modifiers, pressed state, and repeat count

- Updated RawKeyboardSource:
  - RawKeyEvent now carries platform virtualKey + Unicode character + modifiers
  - Keeps platform-specific data isolated at the source layer

- InputManager normalization:
  - Added centralized mapping from platform virtualKey/character -> KeyCode
  - Ensures printable keys resolve to KeyCode::Character
  - Normalizes control keys (Enter, Escape, Tab/Backtab, etc.)
  - Clears character field for non-printable keys
  - Guarantees repeatCount >= 1

Design goals achieved:
- Establishes InputManager as the authoritative normalization boundary
- Prevents platform-specific key codes from leaking into higher layers
- Keeps KeyCode small, durable, and focused on physical/logical keys only
- Prepares clean foundation for future KeyEvent dispatch and CommandMap layers
